Refactor relation analysis interface
