+++ dartagnan/src/main/java/com/dat3m/dartagnan/configuration/OptionNames.java
@@ -57,6 +57,7 @@ public class OptionNames {
     public static final String ENABLE_RELATION_ANALYSIS = "wmm.analysis.relationAnalysis";
     public static final String ENABLE_MUST_SETS = "wmm.analysis.mustSets";
     public static final String ENABLE_EXTENDED_RELATION_ANALYSIS = "wmm.analysis.extendedRelationAnalysis";
+    public static final String RELATION_METHOD = "wmm.analysis.relation";
 
     // Refinement Options
     public static final String BASELINE = "refinement.baseline";
