+++ dartagnan/src/main/java/com/dat3m/dartagnan/wmm/axiom/Acyclicity.java
@@ -82,116 +82,6 @@ public class Acyclicity extends Axiom {
     }
 
     @Override
-    public Map<Relation, RelationAnalysis.ExtendedDelta> computeInitialKnowledgeClosure(
-            Map<Relation, RelationAnalysis.Knowledge> knowledgeMap,
-            Context analysisContext) {
-        long t0 = System.currentTimeMillis();
-        ExecutionAnalysis exec = analysisContext.get(ExecutionAnalysis.class);
-        RelationAnalysis.Knowledge knowledge = knowledgeMap.get(rel);
-        EventGraph may = knowledge.getMaySet();
-        EventGraph must = knowledge.getMustSet();
-        EventGraph newDisabled = new EventGraph();
-        may.apply((e1, e2) -> {
-            if (Tuple.isLoop(e1, e2) || must.contains(e2, e1)) {
-                newDisabled.add(e1, e2);
-            }
-        });
-        Map<Event, List<Event>> mustOut = new HashMap<>();
-        must.apply((e1, e2) -> {
-            if (!Tuple.isLoop(e1, e2)) {
-                mustOut.computeIfAbsent(e1, x -> new ArrayList<>()).add(e2);
-            }
-        });
-        EventGraph current = knowledge.getMustSet();
-        while (true) {
-            EventGraph next = new EventGraph();
-            current.apply((x, y) -> {
-                if (Tuple.isLoop(x, y)) {
-                    boolean implied = exec.isImplied(x, y);
-                    for (Event z : mustOut.getOrDefault(y, List.of())) {
-                        if (!implied && !exec.isImplied(z, y) || exec.areMutuallyExclusive(x, z)) {
-                            continue;
-                        }
-                        if (newDisabled.add(z, x)) {
-                            next.add(x, z);
-                        }
-                    }
-                }
-            });
-            if (next.isEmpty()) {
-                break;
-            }
-            current = next;
-        }
-        newDisabled.retainAll(knowledge.getMaySet());
-        logger.debug("disabled {} edges in {}ms", newDisabled.size(), System.currentTimeMillis() - t0);
-        return Map.of(rel, new RelationAnalysis.ExtendedDelta(newDisabled, EventGraph.empty()));
-    }
-
-    @Override
-    public Map<Relation, RelationAnalysis.ExtendedDelta> computeIncrementalKnowledgeClosure(
-            Relation changed,
-            EventGraph disabled,
-            EventGraph enabled,
-            Map<Relation, RelationAnalysis.Knowledge> knowledgeMap,
-            Context analysisContext) {
-        checkArgument(changed == rel,
-                "misdirected knowledge propagation from relation %s to %s", changed, this);
-        long t0 = System.currentTimeMillis();
-        ExecutionAnalysis exec = analysisContext.get(ExecutionAnalysis.class);
-        RelationAnalysis.Knowledge knowledge = knowledgeMap.get(rel);
-        EventGraph may = knowledge.getMaySet();
-        EventGraph newDisabled = new EventGraph();
-        enabled.apply((e1, e2) -> {
-            if (may.contains(e2, e1)) {
-                newDisabled.add(e2, e1);
-            }
-        });
-        Map<Event, List<Event>> mustIn = new HashMap<>();
-        Map<Event, List<Event>> mustOut = new HashMap<>();
-        knowledge.getMustSet().apply((e1, e2) -> {
-            if (!Tuple.isLoop(e1, e2)) {
-                mustIn.computeIfAbsent(e2, x -> new ArrayList<>()).add(e1);
-                mustOut.computeIfAbsent(e1, x -> new ArrayList<>()).add(e2);
-            }
-        });
-
-        EventGraph current = enabled;
-        while (true) {
-            EventGraph next = new EventGraph();
-            current.apply((x, y) -> {
-                if (!Tuple.isLoop(x, y)) {
-                    boolean implies = exec.isImplied(x, y);
-                    boolean implied = exec.isImplied(y, x);
-                    for (Event w : mustIn.getOrDefault(x, List.of())) {
-                        if (!implied && !exec.isImplied(w, x) || exec.areMutuallyExclusive(w, y)) {
-                            continue;
-                        }
-                        if (newDisabled.add(y, w)) {
-                            next.add(w, y);
-                        }
-                    }
-                    for (Event z : mustOut.getOrDefault(y, List.of())) {
-                        if (!implies && !exec.isImplied(z, y) || exec.areMutuallyExclusive(x, z)) {
-                            continue;
-                        }
-                        if (newDisabled.add(z, x)) {
-                            next.add(x, z);
-                        }
-                    }
-                }
-            });
-            if (next.isEmpty()) {
-                break;
-            }
-            current = next;
-        }
-        newDisabled.retainAll(knowledge.getMaySet());
-        logger.debug("Disabled {} edges in {}ms", newDisabled.size(), System.currentTimeMillis() - t0);
-        return Map.of(rel, new RelationAnalysis.ExtendedDelta(newDisabled, EventGraph.empty()));
-    }
-
-    @Override
     protected EventGraph getEncodeGraph(Context analysisContext) {
         ExecutionAnalysis exec = analysisContext.get(ExecutionAnalysis.class);
         RelationAnalysis ra = analysisContext.get(RelationAnalysis.class);
